+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.8] - 2022-09-05
+### Fixed
+- error when activating addon > copy presets issue
+
+## [0.0.7] - 2022-05-03
+### Fixed
+- error when activating addon > copy presets issue
+
 ## [0.0.6] - 2021-08-06
 ### Changed
 - Preset menu centers like rest of UI in preferences
@@ -35,6 +43,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <!--### Official Rigify Info-->
 
+[0.0.8]:https://github.com/schroef/Save-Custom-Weight-Paint-Presets/releases/tag/v0.0.8
 [0.0.6]:https://github.com/schroef/Save-Custom-Weight-Paint-Presets/releases/tag/v0.0.6
 [0.0.5]:https://github.com/schroef/Save-Custom-Weight-Paint-Presets/releases/tag/v0.0.5
 [0.0.4]:https://github.com/schroef/Save-Custom-Weight-Paint-Presets/releases/tag/v0.0.4